Real-ESRGAN-working-update/requirements_compatible.txt
numpy==2.2.6
opencv-python==4.12.0.88
Pillow>=10.0.0
torch>=2.0.0
tqdm==4.67.1
huggingface-hub==0.34.0
